Implemented CircleCollider2D

+++ source/component/Point2D.h
@@ -5,12 +5,12 @@
 #ifndef POINT2D_H
 #define POINT2D_H
 struct Point2D {
-    double x, y;
+    float x, y;
     Point2D() {
         x = 0;
         y = 0;
     }
-    Point2D(const double x, const double y) {
+    Point2D(const float x, const float y) {
         this->x = x;
         this->y = y;
     }
